update: move the patchownerreference from modelcontroller to kserve and modelmesh
- kserve inject itself as owner to modelcontroller
- modelmesh inject iteslef as owner to modelcontroller
- modelcontroller no need to watch kserve nor modelmesh
- removed duplicated check in dsc for "enabled" which is in the GetManagementStatus
