+++ controllers/components/kserve/kserve_controller_actions.go
@@ -9,8 +9,11 @@ import (
 	operatorv1 "github.com/openshift/api/operator/v1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	k8serr "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/ptr"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	componentsv1 "github.com/opendatahub-io/opendatahub-operator/v2/apis/components/v1"
@@ -256,5 +259,49 @@ func customizeKserveConfigMap(ctx context.Context, rr *odhtypes.ReconciliationRe
 		return err
 	}
 
+	return nil
+}
+
+func patchOwnerReference(ctx context.Context, rr *odhtypes.ReconciliationRequest) error {
+	k, ok := rr.Instance.(*componentsv1.Kserve)
+	if !ok {
+		return fmt.Errorf("resource instance %v is not a componentsv1.Kserve)", rr.Instance)
+	}
+
+	l := logf.FromContext(ctx)
+	mc := &componentsv1.ModelController{}
+	if err := rr.Client.Get(ctx, client.ObjectKey{Name: componentsv1.ModelControllerInstanceName}, mc); err != nil {
+		if k8serr.IsNotFound(err) {
+			return odherrors.NewStopError("ModelController CR not exist: %v", err)
+		}
+		return odherrors.NewStopError("failed to get ModelController CR: %v", err)
+	}
+	for _, owners := range mc.GetOwnerReferences() {
+		if owners.UID == k.GetUID() {
+			return nil // kserve already as owner to modelcontroller, early exit
+		}
+	}
+
+	owners := []metav1.OwnerReference{}
+	owners = append(owners, mc.GetOwnerReferences()...) // keep the existing ones
+	owners = append(owners, metav1.OwnerReference{
+		Kind:               componentsv1.KserveKind,
+		APIVersion:         componentsv1.GroupVersion.String(),
+		Name:               componentsv1.KserveInstanceName,
+		UID:                k.GetUID(),
+		BlockOwnerDeletion: ptr.To(false),
+		Controller:         ptr.To(false),
+	},
+	)
+	mc.SetOwnerReferences(owners)
+	mc.SetManagedFields(nil) // remove managed fields to avoid conflicts when SSA apply
+	opt := &client.PatchOptions{
+		Force:        ptr.To(true),
+		FieldManager: componentsv1.ModelControllerInstanceName,
+	}
+	if err := rr.Client.Patch(ctx, mc, client.Apply, opt); err != nil {
+		return fmt.Errorf("error update ownerreference for CR %s : %w", mc.GetName(), err)
+	}
+	l.Info("Update Ownerreference on kserve change")
 	return nil
 }
