+++ controllers/components/modelmeshserving/modelmeshserving_controller.go
@@ -78,6 +78,7 @@ func (s *componentHandler) NewComponentReconciler(ctx context.Context, mgr ctrl.
 		WithAction(deploy.NewAction(
 			deploy.WithCache(),
 		)).
+		WithAction(patchOwnerReference).
 		WithAction(updatestatus.NewAction()).
 		WithAction(gc.NewAction()).
 		Build(ctx) // include GenerationChangedPredicate no need set in each Owns() above
